pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
